+++ docs/CHANGELOG.md
@@ -3,6 +3,10 @@
 # UNRELEASED
 
 - Update license year, happy 2020!
+- Include cancellation document for document *"CFDI de retención e información de pagos"*.
+- Add `DocumentType` enumerator with keys `cfdi` and `retention` to specify the correct namespace of the request.
+- Testing:
+    - Improve `CancellationTest` to check that `DocumentType` is used correctly.
 
 # Version 1.0.1 2019-10-02